requirement.txt
langchain 
flask 
sentence-transformers 
pypdf 
python-dotenv 
langchain-pinecone 
langchain-openai 
langchain-community 
langchain-google-genai
google-generativeai

